add send_event_to_drive to plugin model for now?
